+++ doc/changelog.md
@@ -78,6 +78,8 @@ go install github.com/magisterquis/curlrevshell@dev
   on port 4444.
 - [`-ctrl-i`](./flags.md#-ctrl-i): Switch from `Ctrl+J` to `Ctrl+S`.
 - [`-debug`](./flags.md#-debug): Debugged DEBUG messages with -debug.
+- [`-log`](./flags.md#-log): Log the TLS fingerprint as well.
+
 - Updated dependencies.
 
 `v0.0.1-beta.7` (2024-10-22)